// generated/orval/fetch/google-health-api/models/observationSampleTime.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Google Health API
 * The Google Health API lets you view and manage health and fitness metrics and measurement data.
 * OpenAPI spec version: v4
 */
import type { CivilDateTime } from './civilDateTime';

/**
 * Represents a sample time of an observed data point.
 */
export interface ObservationSampleTime {
  /** Output only. The civil time in the timezone the subject is in at the time of the observation. */
  readonly civilTime?: CivilDateTime;
  /** Required. The time of the observation. */
  physicalTime?: string;
  /** Required. The offset of the user's local time during the observation relative to the Coordinated Universal Time (UTC). */
  utcOffset?: string;
}